First big download refactoring chunk
